Jun 13, 2025, 1:56 PM

+++ backend/app/infrastructure/http/admin/package_management_routes.py
@@ -29,9 +29,9 @@ class PackageSchema(BaseModel):
     description: Optional[str] = None
     price: int = Field(..., ge=0)
     is_active: bool = True
-    # --- PERUBAHAN 1: Jadikan profile_id opsional ---
     profile_id: Optional[uuid.UUID] = None
-    data_quota_gb: Decimal = Field(..., ge=0, decimal_places=2)
+    # --- PERBAIKAN: Ubah Decimal menjadi float untuk serialisasi JSON ---
+    data_quota_gb: float = Field(..., ge=0)
     duration_days: int = Field(..., gt=0)
     profile: Optional[ProfileSimpleSchema] = None
     model_config = ConfigDict(from_attributes=True)
@@ -62,21 +62,15 @@ def create_package(current_admin: User):
     try:
         package_data = PackageSchema.model_validate(request.get_json())
 
-        # --- PERUBAHAN 2: Logika penetapan profil default ---
-        # Jika frontend tidak mengirim profile_id (karena dropdown sudah dihapus),
-        # cari dan tetapkan profil 'default' secara otomatis.
         if not package_data.profile_id:
             default_profile = db.session.scalar(
                 db.select(PackageProfile).filter_by(profile_name='default')
             )
             if not default_profile:
-                # Ini adalah error server jika profil 'default' tidak ada di DB
                 current_app.logger.critical("Profil 'default' tidak ditemukan di database saat membuat paket!")
                 return jsonify({"message": "Konfigurasi profil sistem error."}), HTTPStatus.INTERNAL_SERVER_ERROR
             
-            # Tetapkan ID dari profil default
             package_data.profile_id = default_profile.id
-        # --- AKHIR PERUBAHAN ---
 
         new_package = Package(**package_data.model_dump(exclude={'id', 'profile'}))
         db.session.add(new_package)
@@ -101,13 +95,12 @@ def update_package(current_admin: User, package_id):
     if not pkg: return jsonify({"message": "Paket tidak ditemukan."}), HTTPStatus.NOT_FOUND
     try:
         package_data = PackageSchema.model_validate(request.get_json())
-        # Jangan izinkan profile_id diubah menjadi null saat update
         update_dict = package_data.model_dump(exclude_unset=True)
         if 'profile_id' in update_dict and update_dict['profile_id'] is None:
             del update_dict['profile_id']
 
         for key, value in update_dict.items():
-            if key not in ['id', 'profile']: # Hindari mengubah ID atau relasi secara langsung
+            if key not in ['id', 'profile']:
                 setattr(pkg, key, value)
         db.session.commit()
         updated_package = db.session.get(Package, pkg.id)
@@ -141,10 +134,8 @@ def get_profiles_for_dropdown(current_admin: User):
     Profil sistem seperti 'default' dan 'expired' akan disembunyikan.
     """
     try:
-        # Buat daftar profil yang tidak ingin ditampilkan di form
         system_profiles_to_hide = ['default', 'expired']
         
-        # Ambil semua profil dari DB, kecuali yang ada di daftar 'hide'
         query = db.select(PackageProfile).where(
             PackageProfile.profile_name.notin_(system_profiles_to_hide)
         ).order_by(PackageProfile.profile_name)
